Add non-root user for running on production
It is safer to run the application under a non-root user on the
production server. Usually it should be safe to run it as root in the
container, but evt. someone figures out how to become root outside the
container as well. Thus using non-root user is the better option.